+++ test/unit/clitests.json
@@ -23,6 +23,7 @@
     "document_spec.js",
     "encodings_spec.js",
     "evaluator_spec.js",
+    "event_utils_spec.js",
     "function_spec.js",
     "message_handler_spec.js",
     "metadata_spec.js",
