+++ web/pdf_history.js
@@ -17,8 +17,8 @@ import {
   isValidRotation,
   parseQueryString,
   PresentationModeState,
-  waitOnEventOrTimeout,
 } from "./ui_utils.js";
+import { waitOnEventOrTimeout } from "./event_utils.js";
 
 // Heuristic value used when force-resetting `this._blockHashChange`.
 const HASH_CHANGE_TIMEOUT = 1000; // milliseconds
